types: omit confusing shared type of options
According to the documentation what was called `AbcParams` (an alias for
`any`) did not really have an overlap between the different methods
where it was used, i.e. it implied a connection where there is none
which can be confusing.

The affected option objects will be individually typed during next
steps.